LayoutTests/imported/mozilla/css-animations/test_document-get-animations-expected.txt


PASS getAnimations for non-animated content 
PASS getAnimations for CSS Animations 
PASS Order of CSS Animations - within an element 
FAIL Order of CSS Animations - across elements assert_equals: Order of second animation returned after tree surgery expected Element node <div style="animation: animLeft 100s"></div> but got Element node <div style="animation: animLeft 100s"></div>
PASS Order of CSS Animations - across and within elements 
FAIL Order of CSS Animations - markup-bound vs free animations assert_equals: getAnimations returns markup-bound and free animations expected 2 but got 1
FAIL Order of CSS Animations - free animations assert_equals: getAnimations returns free animations expected 2 but got 0
FAIL Order of CSS Animations and CSS Transitions assert_equals: Transition comes first expected "[object CSSTransition]" but got "[object CSSAnimation]"
PASS Finished but filling CSS Animations are returned 
PASS Finished but not filling CSS Animations are not returned 
PASS Yet-to-start CSS Animations are returned 
PASS CSS Animations cancelled via the API are not returned 
PASS CSS Animations cancelled and restarted via the API are returned 
FAIL CSS Animations targetting (pseudo-)elements should have correct order after sorting assert_equals: The animation targeting the ::before element comes second expected (string) "::before" but got (undefined) undefined
